requirements.txt
openpyxl>=3.1.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0